refactor(test): move function to helpers
