+++ test/helpers.ts
@@ -1,3 +1,4 @@
+import * as fs from "fs";
 import * as path from "path";
 
 /**
@@ -5,4 +6,14 @@ import * as path from "path";
  */
 export function pathToFixture(...pathArgs: string[]): string {
   return path.join(__dirname, "fixtures", ...pathArgs);
+}
+
+export function loadContentsAsArrayBuffer(filename: string): ArrayBuffer {
+  const buffer = fs.readFileSync(filename);
+  const arr = new ArrayBuffer(buffer.length);
+  const view = new Uint8Array(arr);
+  for (let i = 0; i < buffer.length; i++) {
+    view[i] = buffer[i];
+  }
+  return arr;
 }
